Add files via upload
The project was completed using C++ language. It was a console based game in which the final outcome of the player/game was decided based on certain conditions. This project helped us in developing a better understanding of conditional statements and logics. It also enabled us to use our theoratical knowledge of loops, functions and conditional statements to convert into useful projects.